use long long

+++ dgemm/data.h
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include <time.h>
 
-const int N = 1024;
+const long N = 1024l;
 const double EXPECTED = N * 6.0;
-unsigned long FLOPs = N * N * N * 2l;
+unsigned long long FLOPs = N * N * N * 2l;
 double M1[1024][1024];
 double M2[1024][1024];
 double M3[1024][1024];
